撤稿和恢复功能调试成功

+++ src/main/java/com/wcc/wds/web/utils/FileUtils.java
@@ -1,34 +1,35 @@
-/**  
+/**
  * FileUtils.java
  * com.wcc.wds.web.utils
- * 
+ *
  * @author pengguang
  * @date 2020年10月24日 上午1:18:29
  * 版权所有
  */
 package com.wcc.wds.web.utils;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
+import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.StandardCopyOption;
 
+import com.wcc.wds.web.mapper.WithdrawContributionMapper;
 import com.wcc.wds.web.response.ResponseEnum;
 import com.wcc.wds.web.response.ServiceException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 
 /**
- * TODO(这里用一句话描述这个类的作用)
+ * 通用文件处理
  *
  * @author pengguang
  * @date 2020年10月24日 上午1:18:29
  */
 
 public class FileUtils {
+
+    private static final Logger logger = LoggerFactory.getLogger(FileUtils.class);
+
 
     private static final int numOfEncAndDec = 0x99; // 加密解密秘钥
 
@@ -44,8 +45,12 @@ public class FileUtils {
     public static void encFile(File srcFile, File encFile) {
         if (!encFile.exists()) {
             try {
+                if (!encFile.getParentFile().exists()) {
+                    encFile.getParentFile().mkdirs();
+                }
                 encFile.createNewFile();
             } catch (IOException e) {
+                logger.warn(ResponseEnum.FILE_CREATE_ERROR.msg, e);
                 throw new ServiceException(ResponseEnum.FILE_CREATE_ERROR);
             }
         }
@@ -56,6 +61,7 @@ public class FileUtils {
             }
             fos.flush();
         } catch (Exception e) {
+            logger.warn(ResponseEnum.FILE_ENC_ERROR.msg, e);
             throw new ServiceException(ResponseEnum.FILE_ENC_ERROR);
         }
 
@@ -64,17 +70,25 @@ public class FileUtils {
     /**
      * 文件解密
      *
-     * @param srcFile
+     * @param encFile
      * @param encFile
      * @throws Exception
      */
     public static void decFile(File encFile, File decFile) {
         if (!decFile.exists()) {
             try {
+                if (!decFile.getParentFile().exists()) {
+                    decFile.getParentFile().mkdirs();
+                }
                 decFile.createNewFile();
             } catch (IOException e) {
+                logger.warn(ResponseEnum.FILE_CREATE_ERROR.msg, e);
                 throw new ServiceException(ResponseEnum.FILE_CREATE_ERROR);
             }
+        }
+        if (!encFile.exists()) {
+            logger.warn(ResponseEnum.FILE_DOES_NOT_EXIST_ERROR.msg);
+            throw new ServiceException(ResponseEnum.FILE_DOES_NOT_EXIST_ERROR);
         }
         try (InputStream fis = new FileInputStream(encFile); OutputStream fos = new FileOutputStream(decFile);) {
             while ((dataOfFile = fis.read()) > -1) {
@@ -82,6 +96,7 @@ public class FileUtils {
             }
             fos.flush();
         } catch (Exception e) {
+            logger.warn(ResponseEnum.FILE_DEC_ERROR.msg, e);
             throw new ServiceException(ResponseEnum.FILE_DEC_ERROR);
         }
     }
@@ -96,6 +111,7 @@ public class FileUtils {
         try {
             Files.copy(source.toPath(), target.toPath(), StandardCopyOption.REPLACE_EXISTING);
         } catch (IOException e) {
+            logger.warn(ResponseEnum.FILE_COPY_ERROR.msg, e);
             throw new ServiceException(ResponseEnum.FILE_COPY_ERROR);
         }
     }
@@ -111,7 +127,8 @@ public class FileUtils {
             fileWriter.write(content);
             fileWriter.flush();
         } catch (Exception e) {
-            throw new ServiceException(ResponseEnum.FILE_CREATE_ERROR);
+            logger.warn(ResponseEnum.FILE_SET_CONTENT_ERROR.msg, e);
+            throw new ServiceException(ResponseEnum.FILE_SET_CONTENT_ERROR);
         }
     }
 
@@ -121,9 +138,10 @@ public class FileUtils {
      * @param file
      */
     public static void delete(File file) {
-        try  {
+        try {
             Files.delete(file.toPath());
         } catch (Exception e) {
+            logger.warn(ResponseEnum.FILE_DELETE_ERROR.msg, e);
             throw new ServiceException(ResponseEnum.FILE_DELETE_ERROR);
         }
     }
